Editor: improved navigation between questions
