+++ editor/static/editor/style.css
@@ -48,10 +48,14 @@
 .editor .unit .drag-handle {display:inline-block; position:static; height:25px; width:25px; filter:invert(71%) sepia(55%) saturate(5383%) hue-rotate(169deg) brightness(100%) contrast(94%)}
 .editor .module-thumbnail .drag-handle, .editor .lesson-thumbnail .drag-handle { display:block;position:absolute; left:5px; top:5px; filter:none;}
 
-.editor .lesson-questions .drop-target{ display:inline-block; }
+.editor .lesson-questions .drop-target{ display:inline-block; vertical-align:top;}
 .editor .lesson-questions .drop-target .dock-site { display:none; }
 .editor .lesson-questions .drop-target.drag-over { padding-left:120px; }
-.editor .lesson-questions .dock-space { width:80px; display:inline-block; height:75px;}
-.editor .question-thumbnail { background:#FCFCFC; border:1px solid #DDD; border-radius:5px; margin:5px; padding:0 5px; height:65px; max-width:120px; overflow:hidden;}
+
+.editor .question-thumbnail { background:#FCFCFC; border:1px solid #DDD; border-radius:5px; margin:5px; padding:0 5px; height:65px; max-width:120px; overflow:hidden; cursor:pointer;}
 .editor .question-thumbnail .drag-handle { width:10px; height:10px; display:inline-block; margin-right:5px;}
 .editor .question-thumbnail.selected { background: #f0fafe; border:1px solid #9fddfb; }
+
+.editor .lesson-questions .btn { height: 65px; vertical-align: top; margin: 5px;}
+.editor .lesson-questions .btn-arrow { padding:22px 10px;}
+.editor .lesson-questions .btn-add { padding-top:16px; }
